Fixes bash prompt line wrapping.
All of my colors weren't being properly escaped. It seems that tput
setaf provides the correct escape sequences. All colors (and styles)
must also be enclosed with "\[\]" in order to have them not counted in
the line length of the prompt. That was the crux of the problem. After
fixing .bash_colors the issue was still present. I also needed to
enclose the terminal window title with "\[\]"